pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Banksy FastAPI backend"
package-mode = false
authors = ["Banksy Team"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.9.0"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
